==> 98_download_leitura_2009.R <==
pof::download_pof(2009)
pof::unzip_pof(2009)

# unzip_warn_fail <- function( ... ) tryCatch( { unzip( ... ) } , warning = function( w ) stop( conditionMessage( w ) ) )
library(lodown)
library(tidyverse)

# Leituras baseadas no lodown
# extract the leitura file containing the sas importation instructions
files <- dir(path = "dados/2009/", recursive = T, full.names = T)
leitura <- files[ grep( 'leitura' , tolower( files ) ) ]

z <- instrucoes_sas(leitura)

# write the file back to your second temporary file
tf2 <- tempfile()
writeLines( z , tf2 )

# find each line containing the string `INFILE` or `infile`
all.beginlines <- grep( 'INFILE|infile' , z )

# find line start positions
start.pos <- z[ all.beginlines ] %>% 
  gregexpr(pattern = "\\" , fixed = TRUE) %>% 
  lapply(max) %>% 
  unlist()  %>% 
  magrittr::add(1)

# find line end positions
end.pos <- z[ all.beginlines ] %>% 
  gregexpr(pattern = ".txt") %>% 
  unlist() %>% 
  magrittr::subtract(1)

# isolate the names of all data files to be imported..
data.files.to.import <- substr(z[ all.beginlines ],
                               start.pos, end.pos)

all.file.basenames <- basename(files) %>% 
  strsplit('.' , fixed = TRUE) %>% 
  map_chr(`[[` , 1)

# Depois tirar desse loop e separar cada arquivo em uma função
for ( dfn in data.files.to.import ){
  
  if ( tolower( dfn ) == 't_rendimentos' ) {
    data.file <- files[ which( 't_rendimentos1' == tolower( all.file.basenames ) ) ]
  } else {
    data.file <- files[ which( tolower( dfn ) == tolower( all.file.basenames ) ) ]
  }
  
  curfile <- grep(data.file, pattern = "\\.txt$", value = TRUE)
  
  # figure out which beginline position to use
  cur.beginline <- which( tolower( dfn ) == tolower( data.files.to.import ) )
  
  # import the data file into R
  x <- read_SAScii(
    curfile,
    tf2 ,
    beginline = all.beginlines[ cur.beginline ] ,
    skip_decimal_division = TRUE ,
    sas_encoding = "latin1"
  )
  
  names( x ) <- tolower( names( x ) )
  
  # E ai tem que tirar isso e devolver x
  saveRDS( x , glue::glue("dados/2009/Dados/{tolower(dfn)}.rds"))
}
